+++ CHANGELOG.md
@@ -68,6 +68,11 @@ Remember to align the itemized text with the first line of an item within a list
     positional-only, following deprecation of the keywords in JAX v0.4.21.
   * Non-array arguments to functions in {mod}`jax.lax.linalg` now must be
     specified by keyword. Previously, this raised a DeprecationWarning.
+  * Array-like arguments are now required in several :func:`jax.numpy` APIs,
+    including {func}`~jax.numpy.apply_along_axis`,
+    {func}`~jax.numpy.apply_over_axes`, {func}`~jax.numpy.inner`,
+    {func}`~jax.numpy.outer`, {func}`~jax.numpy.cross`,
+    {func}`~jax.numpy.kron`, and {func}`~jax.numpy.lexsort`.
 
 * Bug fixes
   * {func}`jax.numpy.astype` will now always return a copy when `copy=True`.